Changed: #872 threadid is now in size_t. patch by rti

+++ code/nel/src/misc/common.cpp
@@ -107,12 +107,12 @@ void nlSleep( uint32 ms )
 /*
  * Returns Thread Id (note: on Linux, Process Id is the same as the Thread Id)
  */
-uint getThreadId()
+size_t getThreadId()
 {
 #ifdef NL_OS_WINDOWS
 	return GetCurrentThreadId();
 #elif defined NL_OS_UNIX
-	return uint(pthread_self());
+	return size_t(pthread_self());
 	// doesnt work on linux kernel 2.6	return getpid();
 #endif
